Add modular Settings UI and model management
Add a new Settings module with multiple SwiftUI components and model management UI. New files include AboutSettingsView, AnimatedDownloadingBar, AudioDuckingSettings, DashboardView, GeneralSettingsView, HistorySettingsView, ModelsSettingsView, SettingsSidebar, and SettingsView (registered in the Xcode project). Update the Xcode project (project.pbxproj) to add a Settings group and file references for the new views. Remove the old SettingsView.swift implementation and add temp/ to .gitignore. The ModelsSettingsView contains comprehensive model listing, sorting, NeMo status/install flows, model cards, and storage controls; other views provide UI for general settings, audio ducking, history browsing, dashboard and about info.
