+++ .gitignore
@@ -66,3 +66,4 @@ logs/
 *.tmp
 *.temp
 .cache/
+temp/
